Medical-chatbot-using-LLM-main/requirements.txt
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain-groq
langchain_experimental
-e.